Flask Server (#26)
## What has been done 
The flask server as been set up with the following endpoint :
 - run
 - status
 - snapshot
 - benchmarks
 - benchmark/<bench_id>
 - logs
 - log/log_id
 - results

To enable the server to run the GA a function has been designed to
produce the first generation of individuals.
The exact behavior of each endpoint is described in the documentation.
I did not rewrite it here, as it would make this quite long.

To handle the state of the server, a `Server` class has been created,
this class hold the state of the server and all the methods that produce
the endpoints.

Few functions have been re-designed to improve performance. Some tests have been performed and some early logs created.